fix: cross-review round 3/5 on Batch A
Round 3 findings addressed — 2 bugs + 5 findings, all fixed in-place:

#9 cancel_flow:
- FIX: ff_cancel_flow persists cancellation_policy (Lua was never
  writing it; AlreadyTerminal retry now returns authoritative stored
  value + cancel_reason via HMGET)
- FIX: JoinSet drain logs panics/aborts (was silently swallowed)

#11 index SETs:
- FIX: ff_add_execution_to_flow SADDs flow_index (self-heals flows
  that projector pruned after all-terminal snapshot — new members
  otherwise orphaned with stale summary)
- KEYS 2→3 propagated: Lua + typed wrapper + Server + e2e helper

#13 error kinds:
- FIX: ServerError::is_retryable() + SchedulerError::is_retryable()
  taxonomy: IoError/TryAgain/BusyLoading/ClusterDown/Moved/Ask retryable;
  FatalReceiveError explicitly NOT retryable (response lost)
- FIX: api.rs LibraryLoad arm emits structured kind (was missing)
- FIX: HTTP ErrorBody extended with { kind, retryable } — clients
  parse retry signals from JSON instead of regex on error strings
- REFACTOR: ScriptError moved ff-core → ff-script to eliminate
  ff-core's transitive dep on ferriskey (tokio/rustls/mimalloc).
  ff-core is now pure types; transport layer lives in ff-script.
  ff-sdk import path updated accordingly.

Tests: 66/66 lifecycle + 12/12 API + 44 ff-core + 15 ff-script + 7 sdk.
Zero clippy warnings.
